new source for population info
I've compiled a new JSON for population data that I serve on my own server. The file is the same as the one provided by Geonames, but all populations have been updated to the numbers found on Wiki. Territories that have no permanent residents have a population of 0. And I've changed the key "geonames" to "data".